Add everything

+++ libft_files/ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: akosmeni <akosmeni@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/16 12:19:02 by akosmeni          #+#    #+#             */
-/*   Updated: 2025/05/16 12:25:20 by akosmeni         ###   ########.fr       */
+/*   Updated: 2025/05/27 08:18:23 by akosmeni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	int		len;
-	char	c;
+	size_t		len;
+	char		c;
 
 	if (s)
 	{
